docs_requirements.txt
sphinx
sphinx_rtd_theme
myst_parser
numpydoc
nbsphinx
eesrep
ipython
jupyter
docplex
cplex
mip
pyomo
pandoc